+++ R/count.r
@@ -25,6 +25,8 @@ str_count <- function(string, pattern) {
   switch(type(pattern),
     empty = stri_count_boundaries(string,
       opts_brkiter = stri_opts_brkiter(type = "character")),
+    bound = stri_count_boundaries(string,
+      opts_brkiter = attr(pattern, "options")),
     fixed = stri_count_fixed(string, pattern),
     coll  = stri_count_coll(string, pattern,
       opts_collator = attr(pattern, "options")),