+++ R/extract.r
@@ -18,6 +18,7 @@
 str_extract <- function(string, pattern) {
   switch(type(pattern),
     empty = ,
+    bound = ,
     fixed = stop("Not implemented", call. = FALSE),
     coll  = stri_extract_first_coll(string, pattern,
       opts_collator = attr(pattern, "options")),
@@ -50,6 +51,7 @@ str_extract <- function(string, pattern) {
 str_extract_all <- function(string, pattern, simplify = FALSE) {
   switch(type(pattern),
     empty = ,
+    bound = ,
     fixed = stop("Not implemented", call. = FALSE),
     coll  = stri_extract_all_coll(string, pattern,
       simplify = simplify, omit_no_match = TRUE, attr(pattern, "options")),