+++ R/locate.r
@@ -19,7 +19,8 @@
 #' str_locate(fruit, c("a", "b", "p", "p"))
 str_locate <- function(string, pattern) {
   switch(type(pattern),
-    empty = stop("Not implemented", call. = FALSE),
+    empty = ,
+    bound = stop("Not implemented", call. = FALSE),
     fixed = stri_locate_first_fixed(string, pattern),
     coll  = stri_locate_first_coll(string, pattern,
       opts_collator = attr(pattern, "options")),
@@ -55,6 +56,8 @@ str_locate_all <- function(string, pattern) {
   switch(type(pattern),
     empty = stri_locate_boundaries(string,
       opts_brkiter = stri_opts_brkiter("character")),
+    bound = stri_locate_boundaries(string,
+      opts_brkiter = attr(pattern, "options")),
     fixed = stri_locate_all_fixed(string, pattern, omit_no_match = TRUE),
     regex = stri_locate_all_regex(string, pattern,
       omit_no_match = FALSE, opts_regex = attr(pattern, "options")),